Undid compatibility hunks from the previous patch
The previous patch purposely introduced some off-by-one errors in order to
produce the same exact output as before. THIS patch removes these compatibility
hacks.

+++ src/util3d.h
@@ -64,8 +64,8 @@ typedef vertex GPHUGE * p_vertex;
 /* Maps from normalized space to terminal coordinates */
 #define TERMCOORD_DOUBLE(v,xvar,yvar)           \
 {						\
-    xvar = ((double)(int)((v)->x * xscaler)) + xmiddle;	\
-    yvar = ((double)(int)((v)->y * yscaler)) + ymiddle;	\
+    xvar = (((v)->x * xscaler)) + xmiddle;	\
+    yvar = (((v)->y * yscaler)) + ymiddle;	\
 }
 #define TERMCOORD(v,xvar,yvar)			\
 {						\
